changing os.rename to shutil.move on --format=dir parameter
on os.rename I found two errors:
when dest directory is not empty raises: 'OSError: [Errno 39] Directory not empty'
but with shutil.move it overwrite the directory.

when dest dir is not in the same hd partition of /tmp raises: 'OSError: [Errno 18] Invalid cross-device link'
but with shutil.move it dont occurs.

by Diogo Dutra <dutradda@gmail.com>